+++ morfeusz/FlexionGraph.hpp
@@ -18,7 +18,7 @@ public:
 
     struct Edge {
         InterpretedChunk chunk;
-        unsigned long nextNode;
+        unsigned int nextNode;
     };
 
     void addPath(const std::vector<InterpretedChunk>& path);
